requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
spacy==3.7.2
scikit-learn==1.3.2
langchain==0.0.350
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
httpx==0.25.2 